Load and Prepare Dataset
The dataset has been manually reconstructed to make it easier to understand and integrate. After manually renaming the PDF files in the dataset to the format "patient id_patient name," I began uploading and modifying the dataset on Jupyter Notebook. I chose this environment since it is open-source and contains various important libraries. When the parts of the uploaded dataset are reviewed, it is discovered that it is made up of pdf files with different page numbers, which have been transformed into jpg image files for easier use in this project. During this process, the converted images were saved to the "HTPImages" file and named using the "patient number_patient name_page number" structure. The created dataset has been converted into a structured dataframe, which includes dataset-specific information for clarity and simplicity of understanding. Details like image names and patient numbers have been carefully organised to maintain ethical values and privacy.  Furthermore, a spreadsheet including detailed information and scores for the HTP test drawings was converted, modified, and merged into a final dataframe. This procedure ensured that all relevant data had been gathered into a single, cohesive dataset.