Dataset Updated
After careful consideration, it was decided to label images in the dataset manually. For this purpose, the tool called Label Studio used and objects in images labelled manually. Then, the annotations extracted as COCO format and the problems about the format solved. Finally, the labelled dataset became ready for model training and testing.