Psychological Detection Implemented
In Stage 2 of the project,  the loaded VGG16 model used for feature extraction to better understand the characteristics of the images. With the extracted features, new dataframes created for depression and anxiety, which included the image names, psychological scores (labels), and extracted features. Using this dataframe, machine learning models such as Random Forest, Decision Tree, Logistic Regression, GNB and SVM for psychological label classification task.